One source file per type, no code changes
